Moved JP tests to JP VoidVisitorTest
