--- javaparser-symbol-solver-testing/src/test/resources/issue2366/Test2.java
@@ -1,14 +1,0 @@
-public class Test
-{
-   public class InnerTest
-   {
-       public InnerTest() {}
-   }
-    
-   public Test() {
-   }
-
-   public static void main( String[] args ) { 
-       new Test().new InnerTest();
-   }
-}
